requirements.txt
alpha-vantage==2.1.3
amqp==2.5.2
asgiref==3.2.3
attrs==19.3.0
beautifulsoup4==4.8.2
billiard==3.6.3.0
celery==4.4.1
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0.3
cycler==0.10.0
Django==3.0.4
django-celery-beat==2.0.0
django-cors-headers==3.2.1
django-coverage-plugin==1.8.0
django-crispy-forms==1.9.0
django-debug-toolbar==2.2
django-extensions==2.2.8
django-filter==2.2.0
django-grappelli==2.14.1
django-timezone-field==4.0
django-widget-tweaks==1.4.5
djangorestframework==3.11.0
entrypoints==0.3
factory-boy==2.12.0
Faker==4.0.1
flake8==3.7.9
gunicorn==20.0.4
idna==2.9
importlib-metadata==1.5.0
itypes==1.1.0
Jinja2==2.11.1
kiwisolver==1.1.0
kombu==4.6.8
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.0
mccabe==0.6.1
more-itertools==8.2.0
numpy==1.18.1
packaging==20.3
pandas==1.0.1
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
python-crontab==2.4.0
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
retrying==1.3.3
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
text-unidecode==1.3
tornado==6.0.4
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
wcwidth==0.1.8
zipp==3.1.0
